requirements.txt
keras==2.6.0
numpy==1.19.5
pandas==1.1.1
Pillow==10.2.0
rasterio==1.3.9
scikit_image==0.18.1
scikit_learn==1.0
tensorflow_gpu==2.6.0
tifffile==2024.2.12